Address review feedback
* moves `renderRoot` and all styling features from UpdatingElement to LitElement
* remove unncessary checking for constractable stylesheet support
